requirements.txt
absl-py==0.7.0
networkx==2.3
numpy==1.16.4
scipy==1.3.0
sklearn==0.0
tensorflow==1.14
tqdm==4.32.2
tf_slim
